+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Using records instead of messages
 
+### Fixed
+
+- CSS Styles
+
 ## [2.3.7] - 2024-10-14
 
 ### Added