+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2024-11-24
+
 ### Added
 
 - Script to populate topics
@@ -264,7 +266,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Schema registry support
 
-[Unreleased]: https://github.com/sauljabin/kaskade/compare/v2.3.7...HEAD
+[Unreleased]: https://github.com/sauljabin/kaskade/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/sauljabin/kaskade/compare/v2.3.7...v3.0.0
 [2.3.7]: https://github.com/sauljabin/kaskade/compare/v2.3.6...v2.3.7
 [2.3.6]: https://github.com/sauljabin/kaskade/compare/v2.3.5...v2.3.6
 [2.3.5]: https://github.com/sauljabin/kaskade/compare/v2.3.4...v2.3.5